create benchmark database from github release

+++ load/benchmark_database.py
@@ -23,6 +23,9 @@ FILE_EXTENSION_SCENARIO_SET = "bark_scenarios"
 class BenchmarkDatabase:
     def __init__(self, database_root):
         self.database_root = database_root
+        if not os.path.exists(database_root):
+            logging.error("Given database root does not exist")
+            return
         if database_root.endswith("zip"):
             logging.info("extracting zipped-database {} to temporary directory /tmp/database".format(database_root))
             shutil.rmtree("/tmp/database")
@@ -41,13 +44,16 @@ class BenchmarkDatabase:
                         info_dict = pickle.load(f)
                     self.dataframe = self.dataframe.append(info_dict, ignore_index=True)
         logging.info("The following scenario sets are available")
-        print(self.dataframe.to_string())
+        logging.info("\n"+self.dataframe.to_string()) 
 
     def get_scenario_generator(self, scenario_set_id):
         scenario_generation = ScenarioGeneration()
 
         serialized_file_name = self.dataframe.iloc[scenario_set_id]["Serialized"]
-        serialized_file_path = os.path.join(self.database_root, serialized_file_name)
-        scenario_generation.load_scenario_list(filename=serialized_file_path)
+        if os.path.exists(serialized_file_name):
+            serialized_file_path = serialized_file_name
+        else:
+            serialized_file_path = os.path.join(self.database_root, serialized_file_name)
 
+        scenario_generation.load_scenario_list(filename=serialized_file_path)
         return scenario_generation
